Merged PR 27495: 🎉 feat: set up projects and schema structure (US298657)
The project structure is here. The only BLL class I created is `RoomBLL`, as our only interaction with storing `PlayerModel`s and `TurnModel`s will be in the context of `RoomModel`s; they don't exist without an associated `RoomModel`. If we find we need more BLL functionality, we can add it later as it comes up.

- 🎉 feat(Data): add new `ConnectFour.Data` class library project
- 🎉 feat(Business): add new `ConnectFour.Business` class library project
- ✨ feat(Business): add reference to `ConnectFour.Data`
- ✨ feat(App): add reference to `ConnectFour.Business`
- ✨ feat(Business): add new Model classes
- ✨ feat(Data): add new DTO classes
- ✨ feat(Data): add new Repository classes
- ✨ feat(Business): add new `RoomBLL` class

See #298657

Related work items: #298657, #299216, #299221, #299329, #299333, #299338, #299340